tidy(common): Split off error code (#96)
This is the first part of splitting off some common features from the common library that are used by some, but not all libraries, both new and future. The reason is to better encapsulate functionality and avoid rebuilding the same libraries again.
