+++ src/errors/CMakeLists.txt
@@ -1,0 +1,3 @@
+collect_sources(SFSERR)
+
+shared_add_library(sfserr ${SFSERR_SOURCES})
