+++ src/tools/master_functions.cc
@@ -30,7 +30,7 @@
 #include <sys/statvfs.h>
 
 #include "common/datapack.h"
-#include "common/sfserr.h"
+#include "errors/sfserr.h"
 #include "common/serialization.h"
 #include "common/special_inode_defs.h"
 #include "common/sockets.h"
